requirements.txt
torch
torchvision
Pillow
transformers
easyocr
matplotlib
streamlit
